refactors GeminiLLM to use base class LLM
